Modifications to stamina bar
Thee is  a bug in the stamina bar that it does not get de stamina max value by method, it has to be done by a literal. Later I will search on this bug. (It is a problem because it is only in the build, not in the editor mode).